move brain class from 'apps' to 'jagereye' directory
