+++ jagereye/brain/__init__.py
@@ -1,0 +1,3 @@
+from jagereye.brain.brain import Brain
+
+__all__ = ['Brain']
